# crimson/os/seastore: fix incorrect total space reporting when using multiple shards in RBM